fixed potenial WEIGHT calculation error
On line 63 of main func, we need to amplify the value of log likelihood, since the exp calculation might lead to overflow. I changed the denominator there to 1e20 so it will be more robust. Some other minor refinements were also done, together with refinements earlier today, to make this software more robust.